+++ myMY/customErrorClass.py
@@ -7,4 +7,9 @@ class confirmationError(Exception):
         self.msg = msg
         super().__init__(*args)
 
-
+class noTransactionsError(Exception):
+    """ No transactions record error """
+    msg = str()
+    def __init__(self, *args: object, msg="No transactions record error!") -> None:
+        self.msg = msg
+        super().__init__(*args)